fix(config): tighten additional_per_rank/grace/tank widths
Mirrors the Rust widths: StreamConfig.additional_per_rank map value -> uint8 (streamline.rs additional_streams BTreeMap<String, u8>), GracePeriod.Count -> uint16 (rank.rs u16), HoldingTankConfig.ExpirationDays -> uint8 (placement.rs u8). Out-of-range values now fail at Go YAML unmarshal instead of at the Rust serde boundary.

No omitempty is added to additional_per_rank: Rust StreamConfig.additional_streams has no serde(default), so the field is required on the wire and an empty map must serialize as {} (present), not be dropped. Adds a regression test pinning that presence, two-sided boundary tests for all three fields (inclusive max ok; max+1/overflow/-1 rejected), and repoints the pre-existing type-aware assertions plus two map[string]int test literals to the tightened types.

+++ internal/config/rules_test.go
@@ -977,7 +977,7 @@ func TestValidation_StreamlineAdditionalPerRankMustExist(t *testing.T) {
 				"1": {MinRank: "Associate", Percent: 0.05},
 			},
 			Streams: &StreamConfig{
-				AdditionalPerRank: map[string]int{
+				AdditionalPerRank: map[string]uint8{
 					"Silver":      2,
 					"Nonexistent": 3,
 				},
